Django_Server/requirements.txt
certifi==2017.11.5
chardet==3.0.4
dj-database-url==0.4.2
Django==1.11
djangorestframework==3.6.2
flake8==3.5.0
gunicorn==19.7.1
idna==2.6
mccabe==0.6.1
pew==1.1.0
pipenv==8.3.2
psycopg2==2.7.3.2
pycodestyle==2.3.1
pyflakes==1.6.0
pytz==2017.3
requests==2.18.4
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
